requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
efficientnet_pytorch>=0.7.1
nuscenes-devkit>=1.1.11
pyquaternion>=0.9.9
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.65.0
tensorboard>=2.14.0
